Generalise exports app for other collections
