+++ apps/tools/apps/exports/exports/edition.js
@@ -1,0 +1,50 @@
+const __root = '../../../../..';
+const __src = __root + '/src';
+const __js = __src + '/js';
+const __config = __root + '/config';
+
+const { Members, Permissions } = require(__js + '/database');
+const config = require( __config + '/config.json' );
+
+async function getQuery() {
+	const permission = await Permissions.findOne( { slug: config.permission.member });
+	return {
+		'gocardless.amount': {$gte: 3},
+		permissions: {$elemMatch: {
+			permission,
+			date_expires: {$gte: new Date()}
+		}},
+		delivery_optin: true
+	};
+}
+
+async function getExport(members) {
+	return members
+		.map(member => {
+			const postcode = (member.delivery_address.postcode || '').trim().toUpperCase();
+			return {
+				FirstName: member.firstname,
+				LastName: member.lastname,
+				Address1: member.delivery_address.line1,
+				Address2: member.delivery_address.line2,
+				City: member.delivery_address.city,
+				Postcode: postcode,
+				IsLocal: /^BS[3-9]\D?$/.test(postcode.slice(0, -3)),
+				ReferralLink: member.referralLink
+			};
+		})
+		.sort((a, b) => (
+			(b.IsLocal - a.IsLocal) ||
+				(b.LastName.toLowerCase() > a.LastName.toLowerCase() ? -1 : 1)
+		));
+}
+
+module.exports = {
+	name: 'Edition export',
+	statuses: ['added', 'sent'],
+	collection: Members,
+	itemName: 'members',
+	tableView: 'members',
+	getQuery,
+	getExport
+};
