+++ apps/tools/apps/exports/exports/joinReasons.js
@@ -1,0 +1,32 @@
+const __root = '../../../../..';
+const __src = __root + '/src';
+const __js = __src + '/js';
+
+const { Members } = require(__js + '/database');
+
+async function getQuery() {
+	return {
+		join_reason: {$exists: true, $ne: null}
+	};
+}
+
+async function getExport(members) {
+	return members
+		.map(member => ({
+			Shareable: member.join_shareable ? 'Yes' : 'No',
+			Joined: member.joined,
+			FirstName: member.firstname,
+			Reason: member.join_reason,
+			FirstHeard: member.join_how
+		}));
+}
+
+module.exports = {
+	name: 'Join reasons export',
+	statuses: ['added', 'seen'],
+	collection: Members,
+	itemName: 'join reasons',
+	tableView: 'members',
+	getQuery,
+	getExport
+};
